Rename: mapObject -> tile
