+++ src/game/entities/tile/airGeyserFactory.ts
@@ -6,9 +6,9 @@ import { ColliderComponent, buildCollider } from '@game/components/colliderCompo
 import { RigidBodyComponent } from '@game/components/rigidBodyComponent'
 import { PHYSICS_TAG } from '@game/systems/physicsSystem'
 import { Category, CategorySet } from '../category'
-import { MapObjectFactory } from './mapObjectFactory'
+import { TileEntityFactory } from './mapObjectFactory'
 
-export class AirGeyserFactory extends MapObjectFactory {
+export class AirGeyserFactory extends TileEntityFactory {
   private readonly COLLIDER = {
     type: 'AABB' as const,
     offset: new Vec2(0, 9),
