+++ src/game/entities/tile/dandelionFactory.ts
@@ -1,9 +1,9 @@
 import { Entity } from '@core/ecs/entity'
 import { dandelionAI } from '@game/ai/entity/dandelion/dandelionAI'
 import { AIComponent } from '@game/components/aiComponent'
-import { MapObjectFactory } from './mapObjectFactory'
+import { TileEntityFactory } from './mapObjectFactory'
 
-export class DandelionFactory extends MapObjectFactory {
+export class DandelionFactory extends TileEntityFactory {
   public create(): Entity {
     const entity = super.create()
