+++ src/game/entities/tile/enemy1Factory.ts
@@ -1,6 +1,6 @@
 import { Entity } from '@core/ecs/entity'
 import { Vec2 } from '@core/math/vec2'
-import { slime1AI } from '@game/ai/entity/slime1/slime1AI'
+import { enemy1AI } from '@game/ai/entity/enemy1/enemy1AI'
 import { AIComponent } from '@game/components/aiComponent'
 import { AttackComponent } from '@game/components/attackComponent'
 import { ColliderComponent, buildColliders } from '@game/components/colliderComponent'
@@ -10,9 +10,9 @@ import { RigidBodyComponent } from '@game/components/rigidBodyComponent'
 import { HITBOX_TAG, ATTACK_TAG } from '@game/systems/damageSystem'
 import { PHYSICS_TAG } from '@game/systems/physicsSystem'
 import { Category, CategorySet } from '../category'
-import { MapObjectFactory } from './mapObjectFactory'
+import { TileEntityFactory } from './mapObjectFactory'
 
-export class Slime1Factory extends MapObjectFactory {
+export class Enemy1Factory extends TileEntityFactory {
   private readonly BODY_COLLIDER = {
     type: 'AABB' as const,
     offset: new Vec2(0, 1),
@@ -22,7 +22,8 @@ export class Slime1Factory extends MapObjectFactory {
 
   private readonly HIT_BOX_COLLIDER = {
     type: 'AABB' as const,
-    size: new Vec2(10, 13),
+    size: new Vec2(16, 12),
+    maxClipToTolerance: new Vec2(2, 2),
   }
 
   private readonly RIGID_BODY = {
@@ -33,7 +34,7 @@ export class Slime1Factory extends MapObjectFactory {
   public create(): Entity {
     const entity = super.create()
 
-    entity.addComponent('AI', new AIComponent(slime1AI(entity, this.world)))
+    entity.addComponent('AI', new AIComponent(enemy1AI(entity, this.world)))
     entity.addComponent('RigidBody', new RigidBodyComponent(this.RIGID_BODY))
     entity.addComponent(
       'Collider',
