+++ src/game/entities/tile/kinokoFactory.ts
@@ -13,9 +13,9 @@ import { RigidBodyComponent } from '@game/components/rigidBodyComponent'
 import { AIR_HOLDER_TAG } from '@game/systems/airHolderSystem'
 import { PHYSICS_TAG } from '@game/systems/physicsSystem'
 import { Category, CategorySet } from '../category'
-import { MapObjectFactory } from './mapObjectFactory'
+import { TileEntityFactory } from './mapObjectFactory'
 
-export class KinokoFactory extends MapObjectFactory {
+export class KinokoFactory extends TileEntityFactory {
   private readonly WALL_COLLIDER = {
     type: 'AABB' as const,
     size: new Vec2(20, 64),
