+++ src/game/entities/tile/slime1Factory.ts
@@ -1,22 +1,23 @@
 import { Entity } from '@core/ecs/entity'
 import { Vec2 } from '@core/math/vec2'
-import { SnibeeSetting, snibeeAI } from '@game/ai/entity/snibee/snibeeAI'
+import { slime1AI } from '@game/ai/entity/slime1/slime1AI'
 import { AIComponent } from '@game/components/aiComponent'
 import { AttackComponent } from '@game/components/attackComponent'
 import { ColliderComponent, buildColliders } from '@game/components/colliderComponent'
 import { HorizontalDirectionComponent } from '@game/components/directionComponent'
 import { HPComponent } from '@game/components/hpComponent'
 import { RigidBodyComponent } from '@game/components/rigidBodyComponent'
-import { ATTACK_TAG } from '@game/systems/damageSystem'
+import { HITBOX_TAG, ATTACK_TAG } from '@game/systems/damageSystem'
 import { PHYSICS_TAG } from '@game/systems/physicsSystem'
 import { Category, CategorySet } from '../category'
-import { MapObjectFactory } from './mapObjectFactory'
+import { TileEntityFactory } from './mapObjectFactory'
 
-export class SnibeeFactory extends MapObjectFactory {
+export class Slime1Factory extends TileEntityFactory {
   private readonly BODY_COLLIDER = {
     type: 'AABB' as const,
-    size: new Vec2(10, 13),
-    maxClipToTolerance: new Vec2(SnibeeSetting.maxVelocity / 60, SnibeeSetting.maxVelocity / 60),
+    offset: new Vec2(0, 1),
+    size: new Vec2(16, 12),
+    maxClipToTolerance: new Vec2(2, 2),
   }
 
   private readonly HIT_BOX_COLLIDER = {
@@ -26,13 +27,13 @@ export class SnibeeFactory extends MapObjectFactory {
 
   private readonly RIGID_BODY = {
     mass: 10,
-    airResistance: 0.5,
+    gravityScale: 1,
   }
 
   public create(): Entity {
     const entity = super.create()
 
-    entity.addComponent('AI', new AIComponent(snibeeAI(entity, this.world)))
+    entity.addComponent('AI', new AIComponent(slime1AI(entity, this.world)))
     entity.addComponent('RigidBody', new RigidBodyComponent(this.RIGID_BODY))
     entity.addComponent(
       'Collider',
@@ -49,6 +50,7 @@ export class SnibeeFactory extends MapObjectFactory {
             {
               geometry: this.BODY_COLLIDER,
               category: Category.ENEMY_HITBOX,
+              tag: [HITBOX_TAG],
             },
             {
               geometry: this.HIT_BOX_COLLIDER,
