+++ src/game/entities/tile/snibeeFactory.ts
@@ -1,40 +1,38 @@
 import { Entity } from '@core/ecs/entity'
 import { Vec2 } from '@core/math/vec2'
-import { enemy1AI } from '@game/ai/entity/enemy1/enemy1AI'
+import { SnibeeSetting, snibeeAI } from '@game/ai/entity/snibee/snibeeAI'
 import { AIComponent } from '@game/components/aiComponent'
 import { AttackComponent } from '@game/components/attackComponent'
 import { ColliderComponent, buildColliders } from '@game/components/colliderComponent'
 import { HorizontalDirectionComponent } from '@game/components/directionComponent'
 import { HPComponent } from '@game/components/hpComponent'
 import { RigidBodyComponent } from '@game/components/rigidBodyComponent'
-import { HITBOX_TAG, ATTACK_TAG } from '@game/systems/damageSystem'
+import { ATTACK_TAG } from '@game/systems/damageSystem'
 import { PHYSICS_TAG } from '@game/systems/physicsSystem'
 import { Category, CategorySet } from '../category'
-import { MapObjectFactory } from './mapObjectFactory'
+import { TileEntityFactory } from './mapObjectFactory'
 
-export class Enemy1Factory extends MapObjectFactory {
+export class SnibeeFactory extends TileEntityFactory {
   private readonly BODY_COLLIDER = {
     type: 'AABB' as const,
-    offset: new Vec2(0, 1),
-    size: new Vec2(16, 12),
-    maxClipToTolerance: new Vec2(2, 2),
+    size: new Vec2(10, 13),
+    maxClipToTolerance: new Vec2(SnibeeSetting.maxVelocity / 60, SnibeeSetting.maxVelocity / 60),
   }
 
   private readonly HIT_BOX_COLLIDER = {
     type: 'AABB' as const,
-    size: new Vec2(16, 12),
-    maxClipToTolerance: new Vec2(2, 2),
+    size: new Vec2(10, 13),
   }
 
   private readonly RIGID_BODY = {
     mass: 10,
-    gravityScale: 1,
+    airResistance: 0.5,
   }
 
   public create(): Entity {
     const entity = super.create()
 
-    entity.addComponent('AI', new AIComponent(enemy1AI(entity, this.world)))
+    entity.addComponent('AI', new AIComponent(snibeeAI(entity, this.world)))
     entity.addComponent('RigidBody', new RigidBodyComponent(this.RIGID_BODY))
     entity.addComponent(
       'Collider',
@@ -51,7 +49,6 @@ export class Enemy1Factory extends MapObjectFactory {
             {
               geometry: this.BODY_COLLIDER,
               category: Category.ENEMY_HITBOX,
-              tag: [HITBOX_TAG],
             },
             {
               geometry: this.HIT_BOX_COLLIDER,
