+++ src/game/map/mapBuilder.ts
@@ -5,8 +5,8 @@ import { AirFactory } from '@game/entities/airFactory'
 import { EquipmentTileFactory } from '@game/entities/equipmentTileFactory'
 import { EventSensorFactory } from '@game/entities/eventSensorFactory'
 import { LaserSightFactory } from '@game/entities/laserSightFactory'
-import { AirGeyserFactory } from '@game/entities/mapObject/airGeyserFactory'
-import { PlayerFactory } from '@game/entities/mapObject/playerFactory'
+import { AirGeyserFactory } from '@game/entities/tile/airGeyserFactory'
+import { PlayerFactory } from '@game/entities/tile/playerFactory'
 import { PlayerUIFactory } from '@game/entities/playerUIFactory'
 import { assert } from '@utils/assertion'
 import { TileLayerFactory } from './tileLayerFactory'
